metrics now in long format

+++ instrumentation/src/main/java/tool/DataContainer.java
@@ -4,9 +4,9 @@ package tool;
  * Created by miguel on 22/04/17.
  */
 public class DataContainer {
-    public int bb_blocks = 0;
-    public int methods = 0;
-    public int instructions = 0;
-    public int branch_success = 0;
-    public int branch_fail = 0;
+    public long bb_blocks = 0;
+    public long methods = 0;
+    public long instructions = 0;
+    public long branch_success = 0;
+    public long branch_fail = 0;
 }
